% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{AddMetaData.Assay}
\alias{AddMetaData.Assay}
\alias{AddMetaData}
\title{Add metadata to the exist Assay object}
\usage{
\method{AddMetaData}{Assay}(object, meta.data, col.name = NULL)

AddMetaData(object, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{x}{An object}
}
\description{
Add metadata to the exist Assay object
}
\concept{utils}
