% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{AddTDEseqMetaData.Assay}
\alias{AddTDEseqMetaData.Assay}
\alias{AddTDEseqMetaData}
\title{Add metadata to the exist Assay object}
\usage{
\method{AddTDEseqMetaData}{Assay}(object, meta.data, col.name = NULL)

AddTDEseqMetaData(object, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{x}{An object}
}
\description{
Add metadata to the exist Assay object
}
\concept{utils}
