% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{CalQCMetrics.Assay}
\alias{CalQCMetrics.Assay}
\alias{CalQCMetrics.TDEseq}
\alias{CalQCMetrics}
\title{Calculate meta information}
\usage{
\method{CalQCMetrics}{Assay}(object)

\method{CalQCMetrics}{TDEseq}(object, assay = NULL)

CalQCMetrics(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Calculate meta information
}
