% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{CreateAssayObject}
\alias{CreateAssayObject}
\title{Create an Assay object. Create an Assay object from a feature (e.g. gene) expression matrix. The
expected format of the input matrix is features x cells.}
\usage{
CreateAssayObject(
  counts,
  data = NULL,
  min.cells = 0,
  min.features = 0,
  pct.cells = 0.005,
  min.umi = 10,
  num.core = 1,
  ...
)
}
\arguments{
\item{counts}{Unnormalized data such as raw counts or TPMs}

\item{data}{Prenormalized data; if provided, do not pass \code{counts}}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{...}{Arguments passed to \code{\link{as.sparse}}}
}
\value{
A \code{\link{Assay}} object
}
\description{
Create an Assay object. Create an Assay object from a feature (e.g. gene) expression matrix. The
expected format of the input matrix is features x cells.
}
