% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{CreateTDEseqObject.default}
\alias{CreateTDEseqObject.default}
\alias{CreateTDEseqObject.Assay}
\alias{CreateTDEseqObject.Seurat}
\alias{CreateTDEseqObject.SingleCellExperiment}
\alias{CreateTDEseqObject}
\title{Create a TDEseq object based on an Seurat object}
\usage{
\method{CreateTDEseqObject}{default}(
  counts,
  data = NULL,
  project = "TDEseq",
  assay = "RNA",
  meta.data = NULL,
  species = "hsa",
  names.field = 1,
  names.delim = "_",
  min.cells = 0,
  pct.cells = 0.005,
  min.features = 0,
  min.umi = 10,
  num.core = 1,
  row.names = NULL,
  version = "0.0.1",
  ...
)

\method{CreateTDEseqObject}{Assay}(
  counts,
  project = "TDEseq",
  assay = "RNA",
  names.field = 1,
  names.delim = "_",
  num.core = 1,
  meta.data = NULL,
  ...
)

\method{CreateTDEseqObject}{Seurat}(
  counts,
  project = "TDEseq",
  assay = "RNA",
  min.cells = 0,
  pct.cells = 0.005,
  min.umi = 10,
  num.core = 1,
  min.features = 0,
  row.names = NULL,
  ...
)

\method{CreateTDEseqObject}{SingleCellExperiment}(
  counts,
  project = "TDEseq",
  assay = "RNA",
  slot.use = "counts",
  min.cells = 0,
  pct.cells = 0.005,
  min.umi = 10,
  num.core = 1,
  min.features = 0,
  row.names = NULL,
  ...
)

CreateTDEseqObject(
  counts,
  project = "TDEseq",
  assay = "RNA",
  names.field = 1,
  names.delim = "_",
  ...
)
}
\arguments{
\item{counts}{Either a \code{\link[base]{matrix}}-like object with
unnormalized data with cells as columns and features as rows or an
\code{\link{Assay}}-derived object}

\item{data}{Prenormalized data; if provided, do not pass \code{counts}}

\item{project}{\link{Project} name for the \code{TDEseq} object}

\item{assay}{Name of the initial assay}

\item{names.field}{For the initial identity class for each cell, choose this
field from the cell's name. E.g. If your cells are named as
BARCODE_CLUSTER_CELLTYPE in the input matrix, set \code{names.field} to 3 to
set the initial identities to CELLTYPE.}

\item{names.delim}{For the initial identity class for each cell, choose this
delimiter from the cell's column name. E.g. If your cells are named as
BARCODE-CLUSTER-CELLTYPE, set this to \dQuote{-} to separate the cell name
into its component parts for picking the relevant field.}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{row.names}{When \code{counts} is a \code{data.frame} or
\code{data.frame}-derived object: an optional vector of feature names to be
used}

\item{...}{Arguments passed to other methods}
}
\value{
A \code{\link{TDEseq}} object
}
\description{
Create a TDEseq object based on an Seurat object

Create a TDEseq object based on an SingleCellExperimentobject

Create a \code{TDEseq} object from raw data
}
