% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{DefaultAssay.TDEseq}
\alias{DefaultAssay.TDEseq}
\alias{DefaultAssay<-.TDEseq}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\title{Default Assay}
\usage{
\method{DefaultAssay}{TDEseq}(object, ...)

\method{DefaultAssay}{Assay}(object, ...) <- value

DefaultAssay(object, ...)

DefaultAssay(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
\code{DefaultAssay<-}: An object with the default assay updated
}
\description{
Default Assay
}
