% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{PatternFeature}
\alias{PatternFeature}
\title{PatternHeatmap: Feature plot to show the pattern specific temporal genes. Please first install Seurat, ggplot2 and tidydr packages.}
\usage{
PatternFeature(obj, feature)
}
\arguments{
\item{obj}{The results of TDEseq analysis}

\item{features}{Genes to be shown in feature plot}
}
\description{
PatternHeatmap: Feature plot to show the pattern specific temporal genes. Please first install Seurat, ggplot2 and tidydr packages.
}
\author{
Yue Fan, Shiquan Sun
}
