% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PatternLine}
\alias{PatternLine}
\title{PatternLine: Line plot to show the pattern specific temporal genes. Please first install ggplot2.}
\usage{
PatternLine(obj, feature.show = NULL, cols = NULL)
}
\arguments{
\item{tdeobj}{A tdeseq object}

\item{features}{Genes to be shown in line plot}
}
\description{
PatternLine: Line plot to show the pattern specific temporal genes. Please first install ggplot2.
}
\author{
Yue Fan, Shiquan Sun
}
