% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{tdeseq.default}
\alias{tdeseq.default}
\title{Fitting the constrained spline model to perform temporal differential expression analysis for time course scRNA-seq data}
\usage{
\method{tdeseq}{default}(
  object,
  tde.method = "cell",
  sample.id = NULL,
  stage.id = NULL,
  cov.id = NULL,
  fit.model = "lmm",
  pct = 0.1,
  tde.thr = 0.05,
  lfc = 0,
  max.gcells = Inf,
  min.tcells = 3,
  mod = "FastLMM",
  num.core = 1,
  verbose = FALSE
)
}
\arguments{
\item{object}{Gene expression matrix}

\item{sample.id}{Column names of sample information in metadata.}

\item{stage.id}{Column names of time points/stage information in metadata. The time points or stage information should be provided as numeric data.}

\item{pct}{The percentage of cells where the gene is detected.}

\item{tde.thr}{A numeric value to indicate significant level of DE genes (default = 0.05).}

\item{lfc}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between any two time points. Default is 0.0.}

\item{max.gcells}{Maximum cell per group. If max.gcells is smaller than the given number of cells in each group, the down-sampling will be active.}

\item{min.tcells}{Minimum number of cells in each time points required.}

\item{mod}{Variance component estimation is done either by a linear mixed model implemented in the nlme package(nlme) uniroot, or FastLMM. The uniroot estimation is much more accurate but very time consuming, FastLMM mod is recommended.}

\item{fit.method}{Either 'lmm' or 'lm' to perform (default = lmm).}

\item{tdeseq.method}{Performing 'cell' or 'pseudocell' strategy. Default is 'cell'.}
}
\description{
Fitting the constrained spline model to perform temporal differential expression analysis for time course scRNA-seq data
}
\examples{
data(exampledata)
tde <- CreateTDEseqObject(counts = seurat)
res<-tdeseq(object = tde)

}
