% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.R
\name{tdeseq}
\alias{tdeseq}
\title{Differential expression analysis for time series scRNA-seq data}
\usage{
tdeseq(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Differential expression analysis for time series scRNA-seq data
}
\examples{
tdeseq(object)

}
\concept{data-access}
