% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{tdeseq.TDEseq}
\alias{tdeseq.TDEseq}
\title{TDEseq: detecting temporal expression changes in time-course RNA sequencing studies.}
\usage{
\method{tdeseq}{TDEseq}(
  object,
  assay = "RNA",
  slot.use = "data",
  features = NULL,
  tde.method = "cell",
  tde.param = list(sample.var = "group", stage.var = "stage", fit.model = "lmm", pct =
    0.1, tde.thr = 0.05, lfc = 0, max.gcells = Inf, min.tcells = 3, mod = "FastLMM"),
  num.core = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An TDEseq object}

\item{fit.method}{Either 'lmm' or 'lm' to perform (default = lmm).}

\item{pct}{The percentage of cells where the gene is detected.}

\item{tde.thr}{A numeric value to indicate significant level of DE genes (default = 0.05).}

\item{lfc}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between any two time points. Default is 0.0.}

\item{max.gcells}{Maximum cell per group. If max.gcells is smaller than the given number of cells in each group, the down-sampling will be active.}

\item{min.tcells}{Minimum number of cells in each time points required.}

\item{mod}{Variance component estimation is done either by a linear mixed model implemented in the nlme package or by niroot. The uniroot estimation is much more accurate but very time consuming.}

\item{tdeseq.method}{Performing 'cell' or 'pseudocell' strategy. Default is 'cell'.}
}
\value{
object An TDEseq object
}
\description{
TDEseq: detecting temporal expression changes in time-course RNA sequencing studies.
}
\examples{
data(exampledata)
res <- tdeseq(object=data)

}
\author{
Yue Fan, Shiquan Sun
}
