% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdeseq.object.R
\name{Version.TDEseq}
\alias{Version.TDEseq}
\alias{Version}
\title{Get Version Information}
\usage{
\method{Version}{TDEseq}(object, ...)

Version(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Version Information
}
\examples{
Version(object)

}
\concept{data-access}
