% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{PatternHeatmap}
\alias{PatternHeatmap}
\title{PatternHeatmap: Heatmap to show the pattern specific temporal genes. Please first install Seurat, ggplot2, ComplexHeatmap and circlize packages.}
\usage{
PatternHeatmap(
  obj,
  features = NULL,
  features.show = NULL,
  features.num = 50,
  cols = c("navy", "white", "firebrick3")
)
}
\arguments{
\item{obj}{The results of TDEseq analysis}

\item{features}{Genes to be shown in heatmap}

\item{features.show}{Genes to be annotated in heatmap}

\item{features.num}{Number of genes to be shown for each patterns}

\item{cols}{Color of the heatmap.}
}
\description{
PatternHeatmap: Heatmap to show the pattern specific temporal genes. Please first install Seurat, ggplot2, ComplexHeatmap and circlize packages.
}
\author{
Yue Fan, Shiquan Sun
}
