% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{PatternLine}
\alias{PatternLine}
\title{PatternHeatmap: Line plot to show the pattern specific temporal genes. Please first install ggplot2.}
\usage{
PatternLine(obj, feature.show = NULL, cols = "red")
}
\arguments{
\item{seuobj}{A seurat object with UMAP embeddings been calculated}

\item{features}{Genes to be shown in feature plot}
}
\description{
PatternHeatmap: Line plot to show the pattern specific temporal genes. Please first install ggplot2.
}
\author{
Yue Fan, Shiquan Sun
}
