% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{TDEseq}
\alias{TDEseq}
\title{TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.}
\usage{
TDEseq(
  X,
  meta = NULL,
  z = 0,
  LMM = FALSE,
  pct = 0.1,
  threshold = 0.05,
  logFC_threshold = 0,
  max_cells_per_ident = Inf,
  min_cells_per_timepoints = 0,
  pseudocell = NULL
)
}
\arguments{
\item{X}{The log normalization of expression values, with genes in rows and cells in columns, or Seurat Object.}

\item{meta}{A data frame contained stage index and Group information.}

\item{LMM}{A bool value to indicate whether to perform LMM or LM analysis (default = FALSE).}

\item{pct}{The percentage of cells where the gene is detected.}

\item{threshold}{A numeric value to indicate significant level of DE genes (default = 0.05).}

\item{logFC_threshold}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between any two time points. Default is 0.0.}

\item{max_cells_per_ident}{Down sample each identity class to a max number. Default is no downsampling.}

\item{min_cells_per_timepoints}{Minimum number of cells in one of the time points.}

\item{pseudocell}{Whether perform pseudo cell strategy or not. Default is NULL. Alternatively, users can set pseudocell=20 (or other integer) to perform pseudocell analysis.}
}
\value{
\item{gene}{The name of genes}

\item{pval}{P value for the fixed effect}

\item{padj}{Adjusted P value for the fixed effect}

\item{pattern}{The most appropriate pattern for each DE genes}

\item{ChangePoint}{The change point for DE genes, only exists for peak or recession DE genes}
}
\description{
TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.
}
\examples{
data(exampledata)
data=seurat@assays$RNA@data
meta=seurat@meta.data
res<-TDEseq(data,meta)
}
\author{
Yue Fan, Shiquan Sun
}
