% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{TDEseq}
\alias{TDEseq}
\title{TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.}
\usage{
TDEseq(
  data,
  stage,
  z = 0,
  group = NULL,
  verbose = TRUE,
  LMM = FALSE,
  threshold = 0.05
)
}
\arguments{
\item{data}{The log normalization of expression values, with genes in rows and cells in columns.}

\item{stage}{A vector of stage index or time points.}

\item{group}{A vector to represent the random effect.}

\item{LMM}{A bool value to indicate whether to perform LMM or LM analysis (default = FALSE).}

\item{threshold}{A numeric value to indicate significant level of DE genes (default = 0.05).}
}
\value{
\item{gene}{The name of genes}

\item{pval}{P value for the fixed effect}

\item{padj}{Adjusted P value for the fixed effect}

\item{pattern}{The most appropriate pattern for each DE genes}

\item{ChangePoint}{The change point for DE genes, only exists for peak or recession DE genes}
}
\description{
TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.
}
\author{
Yue Fan, Shiquan Sun
}
