% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDEseq.R
\name{TDEseq}
\alias{TDEseq}
\title{TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.}
\usage{
TDEseq(
  data,
  stage,
  z = 0,
  group = NULL,
  verbose = TRUE,
  LMM = FALSE,
  pct = 0.1,
  threshold = 0.05,
  logFC_threshold = 0,
  max_cells_per_ident = Inf,
  min_cells_per_timepoints = 0,
  pseudocell = NULL
)
}
\arguments{
\item{data}{The log normalization of expression values, with genes in rows and cells in columns.}

\item{stage}{A vector of stage index or time points.}

\item{group}{A vector to represent the random effect.}

\item{LMM}{A bool value to indicate whether to perform LMM or LM analysis (default = FALSE).}

\item{pct}{The percentage of cells where the gene is detected.}

\item{threshold}{A numeric value to indicate significant level of DE genes (default = 0.05).}

\item{logFC_threshold}{Limit testing to genes which show the maximum on average X-fold difference (log-scale) between any two time points. Default is 0.0.}

\item{max_cells_per_ident:}{Down sample each identity class to a max number. Default is no downsampling.}

\item{min_cells_per_timepoints:Minimum}{number of cells in one of the time points.}
}
\value{
\item{gene}{The name of genes}

\item{pval}{P value for the fixed effect}

\item{padj}{Adjusted P value for the fixed effect}

\item{pattern}{The most appropriate pattern for each DE genes}

\item{ChangePoint}{The change point for DE genes, only exists for peak or recession DE genes}
}
\description{
TDEseq: detecting temporal gene expression changes in the developmental stages of single-cell RNA sequencing studies.
}
\examples{
data(exampledata)
data=seurat@assays$RNA@data
stage=seurat@meta.data$stage
res<-TDEseq(data,stage)
}
\author{
Yue Fan, Shiquan Sun
}
